Tests/(Nima) PBA/test #4/pdf_extractor/pj_directory/requirements.txt
# PDF Credit Card Expense Extractor Requirements
# Core dependencies for PDF processing and CLI functionality

# PDF Processing
pdfplumber>=0.9.0          # Primary PDF text extraction library
PyPDF2>=3.0.0              # Backup PDF processing (if needed)

# CLI and User Interface
argparse                   # Command-line argument parsing (built-in)
tabulate>=0.9.0           # Table formatting for CLI output

# Data Processing and Validation
python-dateutil>=2.8.0    # Enhanced date parsing
decimal                   # Precise decimal arithmetic (built-in)

# Configuration Management
PyYAML>=6.0               # YAML configuration file support
configparser              # INI configuration support (built-in)

# Logging and Error Handling
colorama>=0.4.0           # Cross-platform colored terminal output

# Pattern Recognition and Text Processing
regex>=2023.0.0           # Enhanced regular expressions
fuzzywuzzy>=0.18.0        # Fuzzy string matching (optional)
python-Levenshtein>=0.20.0  # Fast string distance calculations

# File and Path Operations
pathlib                   # Modern path handling (built-in)

# Development and Testing Dependencies (optional)
pytest>=7.0.0             # Testing framework
pytest-cov>=4.0.0         # Coverage reporting
black>=23.0.0             # Code formatting
flake8>=6.0.0             # Code linting
mypy>=1.0.0               # Type checking

# System Information (optional)
psutil>=5.9.0             # System and process utilities

# Performance and Optimization (optional)
cachetools>=5.0.0         # Caching utilities

# Documentation (optional)
sphinx>=6.0.0             # Documentation generation
sphinx-rtd-theme>=1.2.0   # Read the Docs theme

# Packaging and Distribution
setuptools>=65.0.0        # Package building
wheel>=0.38.0             # Wheel format support
twine>=4.0.0              # Package uploading

# Version Information
# Python >= 3.8 required